requirements-app.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements-app.in
#
albucore==0.0.24
    # via albumentations
albumentations==2.0.8
    # via -r requirements-app.in
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
click==8.3.2
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
ensure==1.0.4
    # via -r requirements-app.in
fastapi==0.135.3
    # via -r requirements-app.in
filelock==3.25.2
    # via huggingface-hub
flatbuffers==25.12.19
    # via onnxruntime
fsspec==2026.3.0
    # via huggingface-hub
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.9.2
    # via -r requirements-app.in
idna==3.11
    # via
    #   anyio
    #   httpx
joblib==1.5.3
    # via -r requirements-app.in
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.4
    # via onnx
mpmath==1.3.0
    # via sympy
numpy==2.4.4
    # via
    #   -r requirements-app.in
    #   albucore
    #   albumentations
    #   ml-dtypes
    #   onnx
    #   onnxruntime
    #   opencv-python-headless
    #   scipy
onnx==1.21.0
    # via -r requirements-app.in
onnxruntime==1.24.4
    # via -r requirements-app.in
opencv-python-headless==4.13.0.92
    # via
    #   -r requirements-app.in
    #   albucore
    #   albumentations
packaging==26.0
    # via
    #   huggingface-hub
    #   onnxruntime
pillow==12.2.0
    # via -r requirements-app.in
prometheus-client==0.24.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.1.0
    # via -r requirements-app.in
protobuf==7.34.1
    # via
    #   onnx
    #   onnxruntime
psutil==7.2.2
    # via -r requirements-app.in
pydantic==2.12.5
    # via
    #   albumentations
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.20.0
    # via rich
python-box==7.4.1
    # via -r requirements-app.in
python-dotenv==1.2.2
    # via -r requirements-app.in
python-multipart==0.0.24
    # via -r requirements-app.in
pyyaml==6.0.3
    # via
    #   -r requirements-app.in
    #   albumentations
    #   huggingface-hub
rich==14.3.3
    # via typer
scipy==1.17.1
    # via albumentations
shellingham==1.5.4
    # via typer
simsimd==6.5.16
    # via albucore
six==1.17.0
    # via ensure
starlette==0.52.1
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
stringzilla==4.6.0
    # via albucore
sympy==1.14.0
    # via onnxruntime
tqdm==4.67.3
    # via huggingface-hub
typer==0.24.1
    # via huggingface-hub
types-pyyaml==6.0.12.20260408
    # via -r requirements-app.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   onnx
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.44.0
    # via -r requirements-app.in
